pyproject.toml
[project]
name = "temporary-langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "openai",
    "tiktoken",
    "langchain-community",
    "langchain-openai",
    "langchain-chroma",
    "haystack-ai>=2.14.2",
    "sentence-transformers"
]
